Reorganize reflect serializer models
